modified amount and price fields in invoice items to double

+++ omods/core/api/src/main/java/org/openmrs/module/icare/store/models/StockInvoiceItem.java
@@ -15,129 +15,127 @@ import java.util.HashMap;
 import java.util.Map;
 
 @Entity
-@Table(name="st_stock_invoice_item")
+@Table(name = "st_stock_invoice_item")
 public class StockInvoiceItem extends BaseOpenmrsData implements java.io.Serializable, JSONConverter {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "stock_invoice_item_id",unique= true, nullable = false)
-    private Integer id;
-
-    @ManyToOne
-    @JoinColumn(name = "stock_invoice_id")
-    private StockInvoice stockInvoice;
-
-    @ManyToOne
-    @JoinColumn(name = "item_id")
-    private Item item;
-
-    @Column(name = "batch_no", length=30)
-    private String batchNo;
-
-    @Column(name = "order_quantity")
-    private Integer orderQuantity;
-
-    @Column(name = "batch_quantity")
-    private Integer batchQuantity;
-
-    @Column(name = "expiry_date")
-    private Date expiryDate;
-
-
-    @ManyToOne
-    @JoinColumn(name = "uom")
-    private Concept uom;
-
-    @Column(name = "unit_price")
-    private Integer unitPrice;
-
-    @Column(name = "amount")
-    private Integer amount;
-
-
-    @Override
-    public Integer getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(Integer id) {
-        this.id =id;
-    }
-
-    public Item getItem() {
-        return item;
-    }
-
-    public void setItem(Item item) {
-        this.item = item;
-    }
-
-    public String getBatchNo() {
-        return batchNo;
-    }
-
-    public void setBatchNo(String batchNo) {
-        this.batchNo = batchNo;
-    }
-
-    public Integer getOrderQuantity() {
-        return orderQuantity;
-    }
-
-    public void setOrderQuantity(Integer orderQuantity) {
-        this.orderQuantity = orderQuantity;
-    }
-
-    public Integer getBatchQuantity() {
-        return batchQuantity;
-    }
-
-    public void setBatchQuantity(Integer batchQuantity) {
-        this.batchQuantity = batchQuantity;
-    }
-
-    public Date getExpiryDate() {
-        return expiryDate;
-    }
-
-    public void setExpiryDate(Date expiryDate) {
-        this.expiryDate = expiryDate;
-    }
-
-    public Concept getUom() {
-        return uom;
-    }
-
-    public void setUom(Concept uom) {
-        this.uom = uom;
-    }
-
-    public Integer getUnitPrice() {
-        return unitPrice;
-    }
-
-    public void setUnitPrice(Integer unitPrice) {
-        this.unitPrice = unitPrice;
-    }
-
-    public Integer getAmount() {
-        return amount;
-    }
-
-    public void setAmount(Integer amount) {
-        this.amount = amount;
-    }
-
-    public void setStockInvoice(StockInvoice stockInvoice) {
-        this.stockInvoice = stockInvoice;
-    }
-
-    public StockInvoice getStockInvoice() {
-        return stockInvoice;
-    }
-
-    @Override
+	
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "stock_invoice_item_id", unique = true, nullable = false)
+	private Integer id;
+	
+	@ManyToOne
+	@JoinColumn(name = "stock_invoice_id")
+	private StockInvoice stockInvoice;
+	
+	@ManyToOne
+	@JoinColumn(name = "item_id")
+	private Item item;
+	
+	@Column(name = "batch_no", length = 30)
+	private String batchNo;
+	
+	@Column(name = "order_quantity")
+	private Integer orderQuantity;
+	
+	@Column(name = "batch_quantity")
+	private Integer batchQuantity;
+	
+	@Column(name = "expiry_date")
+	private Date expiryDate;
+	
+	@ManyToOne
+	@JoinColumn(name = "uom")
+	private Concept uom;
+	
+	@Column(name = "unit_price")
+	private Double unitPrice;
+	
+	@Column(name = "amount")
+	private Double amount;
+	
+	@Override
+	public Integer getId() {
+		return id;
+	}
+	
+	@Override
+	public void setId(Integer id) {
+		this.id = id;
+	}
+	
+	public Item getItem() {
+		return item;
+	}
+	
+	public void setItem(Item item) {
+		this.item = item;
+	}
+	
+	public String getBatchNo() {
+		return batchNo;
+	}
+	
+	public void setBatchNo(String batchNo) {
+		this.batchNo = batchNo;
+	}
+	
+	public Integer getOrderQuantity() {
+		return orderQuantity;
+	}
+	
+	public void setOrderQuantity(Integer orderQuantity) {
+		this.orderQuantity = orderQuantity;
+	}
+	
+	public Integer getBatchQuantity() {
+		return batchQuantity;
+	}
+	
+	public void setBatchQuantity(Integer batchQuantity) {
+		this.batchQuantity = batchQuantity;
+	}
+	
+	public Date getExpiryDate() {
+		return expiryDate;
+	}
+	
+	public void setExpiryDate(Date expiryDate) {
+		this.expiryDate = expiryDate;
+	}
+	
+	public Concept getUom() {
+		return uom;
+	}
+	
+	public void setUom(Concept uom) {
+		this.uom = uom;
+	}
+	
+	public Double getUnitPrice() {
+		return unitPrice;
+	}
+	
+	public void setUnitPrice(Double unitPrice) {
+		this.unitPrice = unitPrice;
+	}
+	
+	public Double getAmount() {
+		return amount;
+	}
+	
+	public void setAmount(Double amount) {
+		this.amount = amount;
+	}
+	
+	public void setStockInvoice(StockInvoice stockInvoice) {
+		this.stockInvoice = stockInvoice;
+	}
+	
+	public StockInvoice getStockInvoice() {
+		return stockInvoice;
+	}
+	
+	@Override
     public Map<String, Object> toMap() {
 
         HashMap<String,Object> invoiceItemObject = new HashMap<>();
@@ -164,33 +162,34 @@ public class StockInvoiceItem extends BaseOpenmrsData implements java.io.Seriali
         invoiceItemObject.put("uom",uomObject);
         return invoiceItemObject;
     }
-
-    public static StockInvoiceItem fromMap(Map<String,Object> stockInvoiceItemMap) throws ParseException {
-
-        StockInvoiceItem stockInvoiceItem = new StockInvoiceItem();
-
-        Item item = new Item();
-        item.setUuid(((Map)stockInvoiceItemMap.get("item")).get("uuid").toString());
-        stockInvoiceItem.setItem(item);
-
-        Concept uom = new Concept();
-        uom.setUuid(((Map) stockInvoiceItemMap.get("uom")).get("uuid").toString());
-        stockInvoiceItem.setUom(uom);
-
-        stockInvoiceItem.setBatchNo(stockInvoiceItemMap.get("batchNo").toString());
-        stockInvoiceItem.setUnitPrice((Integer) stockInvoiceItemMap.get("unitPrice"));
-        stockInvoiceItem.setAmount((Integer) stockInvoiceItemMap.get("amount"));
-        stockInvoiceItem.setBatchQuantity((Integer) stockInvoiceItemMap.get("batchQuantity"));
-        stockInvoiceItem.setOrderQuantity((Integer) stockInvoiceItemMap.get("orderQuantity"));
-
-        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        if (stockInvoiceItemMap.get("expiryDate").toString().length() == 10) {
-            stockInvoiceItem.setExpiryDate(dateFormat.parse(stockInvoiceItemMap.get("expiryDate").toString()));
-        } else {
-            stockInvoiceItem.setExpiryDate(dateFormat.parse(stockInvoiceItemMap.get("expiryDate").toString().substring(0, stockInvoiceItemMap.get("expiryDate").toString().indexOf("T"))));
-        }
-
-        return stockInvoiceItem;
-
-    }
+	
+	public static StockInvoiceItem fromMap(Map<String, Object> stockInvoiceItemMap) throws ParseException {
+		
+		StockInvoiceItem stockInvoiceItem = new StockInvoiceItem();
+		
+		Item item = new Item();
+		item.setUuid(((Map) stockInvoiceItemMap.get("item")).get("uuid").toString());
+		stockInvoiceItem.setItem(item);
+		
+		Concept uom = new Concept();
+		uom.setUuid(((Map) stockInvoiceItemMap.get("uom")).get("uuid").toString());
+		stockInvoiceItem.setUom(uom);
+		
+		stockInvoiceItem.setBatchNo(stockInvoiceItemMap.get("batchNo").toString());
+		stockInvoiceItem.setUnitPrice((Double) stockInvoiceItemMap.get("unitPrice"));
+		stockInvoiceItem.setAmount((Double) stockInvoiceItemMap.get("amount"));
+		stockInvoiceItem.setBatchQuantity((Integer) stockInvoiceItemMap.get("batchQuantity"));
+		stockInvoiceItem.setOrderQuantity((Integer) stockInvoiceItemMap.get("orderQuantity"));
+		
+		SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+		if (stockInvoiceItemMap.get("expiryDate").toString().length() == 10) {
+			stockInvoiceItem.setExpiryDate(dateFormat.parse(stockInvoiceItemMap.get("expiryDate").toString()));
+		} else {
+			stockInvoiceItem.setExpiryDate(dateFormat.parse(stockInvoiceItemMap.get("expiryDate").toString()
+			        .substring(0, stockInvoiceItemMap.get("expiryDate").toString().indexOf("T"))));
+		}
+		
+		return stockInvoiceItem;
+		
+	}
 }
